Fix balloon generator
